requirements.txt
discord.py==1.2.4
peewee==3.11.2
